Validation work

+++ database/migrations/20210630125121_datahistoryjob_conversions/postgres.sql
@@ -2,7 +2,7 @@
 ALTER TABLE datahistoryjob
     ADD conversion_interval DOUBLE PRECISION,
     ADD overwrite_data boolean,
-    ADD decimal_place_comparison NUMERIC;
+    ADD decimal_place_comparison INTEGER;
 -- +goose Down
 ALTER TABLE datahistoryjob
     DROP conversion_interval,
